refactor: extract ScenarioParameterBinder so workflow catalog can reuse it
This extracts the shared parameter-binding logic from ScenarioCatalog into
ScenarioParameterBinder, allowing WorkflowScenarioCatalog to reuse the same
tested implementation of override coercion, Min/Max validation, and default-value
fallback instead of reimplementing it.
